Finished fsm homework adding double jump and crouching. Prepare for sprite homework

+++ world/world.h
@@ -14,7 +14,7 @@ public:
     void add_platform(float x, float y, float width, float height);
     bool collides(const Vec<float>& position) const;
     GameObject* create_player();
-    void move_to(Vec<float>& position, const Vec<int>& size, Vec<float>& velocity);
+    void move_to(Vec<float>& position, const Vec<float>& size, Vec<float>& velocity);
     void update(float dt);
 
     Tilemap tilemap;
